Fixed many type-size-related warnings.

+++ tests/unit/src/receiver/zmq/zmq_receiver.cpp
@@ -207,7 +207,7 @@ namespace APSITests
                 uint32_t bundle_idx = loc / get_params()->items_per_bundle();
                 uint32_t bundle_offset = loc - bundle_idx * get_params()->items_per_bundle();
                 fill_n(
-                    rp_vec.begin() + bundle_offset * get_params()->item_params().felts_per_item,
+                    rp_vec.begin() + static_cast<size_t>(bundle_offset) * get_params()->item_params().felts_per_item,
                     get_params()->item_params().felts_per_item, 0);
             }
